test/pyproject.toml
[project]
name = "data-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.31.0",
    "Faker>=25.2.0",
    "rich>=13.7.1",
    "typer>=0.12.3",
    "pyjwt>=2.10.1",
]
